Move remaining js files to typescript  - jest  - profiles  - getBaseName
