+++ jest.config.js
@@ -1,0 +1,18 @@
+module.exports = {
+    preset: 'ts-jest',
+    testEnvironment: 'node',
+    coverageDirectory: './coverage',
+    collectCoverageFrom: [
+        'src/**/*.js',
+        '!src/**/stories/*'
+    ],
+    setupFiles: [
+        '<rootDir>/config/setupTests.ts'
+    ],
+    roots: [
+        '<rootDir>/src/'
+    ],
+    moduleNameMapper: {
+        '\\.(css|scss)$': 'identity-obj-proxy'
+    }
+};
