+++ package.json
@@ -2,23 +2,6 @@
   "name": "custom-policies-ui-front",
   "version": "1.1.0",
   "private": false,
-  "jest": {
-    "coverageDirectory": "./coverage/",
-    "collectCoverage": true,
-    "collectCoverageFrom": [
-      "src/**/*.js",
-      "!src/**/stories/*"
-    ],
-    "setupFiles": [
-      "<rootDir>/config/setupTests.js"
-    ],
-    "roots": [
-      "<rootDir>/src/"
-    ],
-    "moduleNameMapper": {
-      "\\.(css|scss)$": "identity-obj-proxy"
-    }
-  },
   "dependencies": {
     "@babel/runtime": "^7.7.2",
     "@patternfly/patternfly": "^2.40.9",
@@ -50,6 +33,7 @@
     "@babel/preset-react": "^7.7.0",
     "@redhat-cloud-services/frontend-components-config": "0.0.10",
     "@types/axios": "^0.14.0",
+    "@types/jest": "^24.0.23",
     "@types/node": "^12.12.14",
     "@types/react": "^16.9.11",
     "@types/react-dom": "^16.9.4",
@@ -82,6 +66,7 @@
     "stylelint-config-recommended-scss": "^4.0.0",
     "stylelint-scss": "^3.12.1",
     "ts-essentials": "^3.0.5",
+    "ts-jest": "^24.2.0",
     "ts-node": "^8.5.4",
     "typescript": "^3.7.2",
     "webpack": "^4.41.2",
@@ -102,7 +87,7 @@
     "deploy": "npm-run-all build:prod lint test",
     "verify": "npm-run-all build lint test",
     "nightly": "npm run deploy",
-    "proxy": "SPANDX_CONFIG=\"./profiles/local-frontend-and-api.js\" bash $PROXY_PATH/scripts/run.sh"
+    "proxy": "SPANDX_CONFIG=\"profiles/local-frontend-and-api.ts\" bash $PROXY_PATH/scripts/run.sh"
   },
   "insights": {
     "appname": "custom-policies"
