+++ src/entry-dev.js
@@ -5,7 +5,7 @@ import { Provider } from 'react-redux';
 import { init } from './store';
 import App from './app/App';
 import logger from 'redux-logger';
-import getBaseName from './Utilities/getBaseName';
+import getBaseName from './utils/getBaseName';
 
 ReactDOM.render(
     <Provider store={ init(logger).getStore() }>
