+++ src/entry.tsx
@@ -4,7 +4,7 @@ import { BrowserRouter as Router } from 'react-router-dom';
 import { Provider } from 'react-redux';
 import { init } from './store';
 import App from './app/App';
-import getBaseName from './Utilities/getBaseName';
+import getBaseName from './utils/getBaseName';
 
 ReactDOM.render(
     <Provider store={ init().getStore() }>
